+++ tests/Unit/Models/ReminderTest.php
@@ -133,6 +133,29 @@ class ReminderTest extends TestCase
     }
 
     /** @test */
+    public function it_calculates_next_expected_date_in_timezone()
+    {
+        config(['app.timezone' => 'Europe/Paris']);
+
+        $reminder = new Reminder;
+        $reminder->initial_date = '1980-05-01';
+        $reminder->frequency_type = 'year';
+        $reminder->frequency_number = 1;
+
+        Carbon::setTestNow(Carbon::create(2000, 4, 30, 21, 59, 59));
+        $this->assertEquals(
+            '2000-05-01',
+            $reminder->calculateNextExpectedDateOnTimezone()->toDateString()
+        );
+
+        Carbon::setTestNow(Carbon::create(2000, 4, 30, 22, 00, 00));
+        $this->assertEquals(
+            '2001-05-01',
+            $reminder->calculateNextExpectedDateOnTimezone()->toDateString()
+        );
+    }
+
+    /** @test */
     public function it_schedules_a_reminder_for_one_user()
     {
         Carbon::setTestNow(Carbon::create(2017, 2, 1));
